pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evalview"
version = "0.3.2"
description = "Open-source testing and regression detection framework for AI agents. Golden baseline diffing, CI/CD integration, works with LangGraph, CrewAI, OpenAI, Anthropic Claude, HuggingFace, Ollama, and MCP."
readme = "README.md"
license = "Apache-2.0"
authors = [{name = "EvalView Team", email = "hidai@evalview.com"}]
requires-python = ">=3.9"
keywords = [
    "ai", "agents", "testing", "evaluation", "llm", "langchain", "langgraph", "crewai", "openai",
    "anthropic", "claude", "huggingface", "ollama", "mcp",
    "multi-agent", "pytest-ai", "ai-agent-testing", "llm-testing", "agent-evaluation",
    "regression-testing", "golden-baseline", "ci-cd-testing", "yaml-testing", "tool-calling",
    "agent-regression", "llm-regression", "skill-testing", "mcp-contract-testing",
    "pass-at-k", "non-deterministic-testing", "ai-ci-cd"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "openai>=1.12.0",
    "anthropic>=0.39.0",
    "rich>=13.7.0",
    "prompt_toolkit>=3.0.0",
    "httpx>=0.26.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# HTML reports with interactive charts
reports = [
    "jinja2>=3.0",
    "plotly>=5.0",
]
# Watch mode for development
watch = [
    "watchdog>=3.0",
]
# Telemetry (opt-out anonymous usage analytics)
telemetry = [
    "posthog>=3.0.0",
]
# All optional features
all = [
    "jinja2>=3.0",
    "plotly>=5.0",
    "watchdog>=3.0",
    "posthog>=3.0.0",
]
# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black==24.10.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "jinja2>=3.0",
    "plotly>=5.0",
    "watchdog>=3.0",
    "anthropic>=0.39.0",
    # For mock agent in dogfood tests
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
evalview = "evalview.cli:main"

[project.entry-points."pytest11"]
evalview = "evalview.pytest_plugin"

[project.urls]
Homepage = "https://github.com/hidai25/eval-view"
Documentation = "https://github.com/hidai25/eval-view#readme"
Repository = "https://github.com/hidai25/eval-view.git"
"Issue Tracker" = "https://github.com/hidai25/eval-view/issues"
Changelog = "https://github.com/hidai25/eval-view/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["evalview*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
check_untyped_defs = true

# Exclude CLI which has many click decorators that are hard to type
[[tool.mypy.overrides]]
module = "evalview.cli"
ignore_errors = true

[[tool.mypy.overrides]]
module = "evalview.reporters.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "evalview.tracking.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "evalview.recorder"
ignore_errors = true

[[tool.mypy.overrides]]
module = "evalview.evaluators.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "evalview.core.loader"
ignore_errors = true

[[tool.mypy.overrides]]
module = "evalview.adapters.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "evalview.telemetry.*"
ignore_errors = true
